+++ packages/ts-components/src/components/article-flag/ArticleFlag.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { colours } from '@times-components/styleguide';
+import { colours } from '@times-components/ts-styleguide';
 import { LiveArticleFlag, BreakingArticleFlag } from './LiveArticleFlag';
 import {
   ArticleFlagContainer,