+++ packages/article-extras/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.46](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.10.45...@times-components/article-extras@0.10.46) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
 ## [0.10.45](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.10.44...@times-components/article-extras@0.10.45) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/article-extras