+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.18.25](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.18.24...@times-components/article-summary@3.18.25) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/article-summary
+
+
+
+
+
 ## [3.18.24](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.18.23...@times-components/article-summary@3.18.24) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/article-summary