+++ packages/caption/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.171](https://github.com/newsuk/times-components/compare/@times-components/caption@3.0.170...@times-components/caption@3.0.171) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/caption
+
+
+
+
+
 ## [3.0.170](https://github.com/newsuk/times-components/compare/@times-components/caption@3.0.169...@times-components/caption@3.0.170) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/caption