+++ packages/date-publication/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.22.21](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.22.20...@times-components/date-publication@0.22.21) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/date-publication
+
+
+
+
+
 ## [0.22.20](https://github.com/newsuk/times-components/compare/@times-components/date-publication@0.22.19...@times-components/date-publication@0.22.20) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/date-publication