+++ packages/message-bar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.20](https://github.com/newsuk/times-components/compare/@times-components/message-bar@0.4.19...@times-components/message-bar@0.4.20) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/message-bar
+
+
+
+
+
 ## [0.4.19](https://github.com/newsuk/times-components/compare/@times-components/message-bar@0.4.18...@times-components/message-bar@0.4.19) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/message-bar