+++ packages/save-and-share-bar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.110](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.9.109...@times-components/save-and-share-bar@0.9.110) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/save-and-share-bar
+
+
+
+
+
 ## [0.9.109](https://github.com/newsuk/times-components/compare/@times-components/save-and-share-bar@0.9.108...@times-components/save-and-share-bar@0.9.109) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/save-and-share-bar