+++ packages/save-star-web/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.106](https://github.com/newsuk/times-components/compare/@times-components/save-star-web@0.6.105...@times-components/save-star-web@0.6.106) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/save-star-web
+
+
+
+
+
 ## [0.6.105](https://github.com/newsuk/times-components/compare/@times-components/save-star-web@0.6.104...@times-components/save-star-web@0.6.105) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/save-star-web