+++ packages/ssr/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.31.6](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.31.5...@times-components/ssr@2.31.6) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/ssr
+
+
+
+
+
 ## [2.31.5](https://github.com/newsuk/times-components/compare/@times-components/ssr@2.31.4...@times-components/ssr@2.31.5) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/ssr