+++ packages/sticky/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.33](https://github.com/newsuk/times-components/compare/@times-components/sticky@0.3.32...@times-components/sticky@0.3.33) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/sticky
+
+
+
+
+
 ## [0.3.32](https://github.com/newsuk/times-components/compare/@times-components/sticky@0.3.31...@times-components/sticky@0.3.32) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/sticky