+++ packages/ts-styleguide/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.35.2](https://github.com/newsuk/times-components/compare/@times-components/ts-styleguide@1.35.1...@times-components/ts-styleguide@1.35.2) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/ts-styleguide
+
+
+
+
+
 ## 1.35.1 (2022-05-11)
 
 **Note:** Version bump only for package @times-components/ts-styleguide