+++ packages/video/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.11.43](https://github.com/newsuk/times-components/compare/@times-components/video@4.11.42...@times-components/video@4.11.43) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/video
+
+
+
+
+
 ## [4.11.42](https://github.com/newsuk/times-components/compare/@times-components/video@4.11.41...@times-components/video@4.11.42) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/video