+++ packages/watermark/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.82](https://github.com/newsuk/times-components/compare/@times-components/watermark@2.7.81...@times-components/watermark@2.7.82) (2022-05-12)
+
+**Note:** Version bump only for package @times-components/watermark
+
+
+
+
+
 ## [2.7.81](https://github.com/newsuk/times-components/compare/@times-components/watermark@2.7.80...@times-components/watermark@2.7.81) (2022-05-11)
 
 **Note:** Version bump only for package @times-components/watermark